+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The changelog is currently handwritten.
 - Updated dependencies.
   - `console.log` is now printed to the terminal in unit tests again. This [broke with `karma@1.5.0`](https://github.com/karma-runner/karma/issues/2602).
   - Update to new version of `awesome-typescript-loader`. Now you **don't need** to set `declaration: true` in your `tsconfig.json` for browser components. We do this automatically.
+- We throw an error if typings cannot be found after build.
 
 # 1.0.1-61 (2017-02-23)